backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
psycopg2-binary
beautifulsoup4
google-generativeai
python-multipart
requests
python-dotenv
pydantic
